chore(api): move files under relevant directories
