+++ src/index.ts
@@ -1,7 +1,7 @@
-import createEffect from "./createEffect";
-import createElement from "./createElement";
-import createFragment from "./createFragment";
-import createState from "./createState";
+import createEffect from "./api/state/createEffect";
+import createElement from "./api/runtime/createElement";
+import createFragment from "./api/runtime/createFragment";
+import createState from "./api/state/createState";
 
 export default {
   createElement,
